chore: add a single canonical type for MockServer
