+++ packages/hoppscotch-common/src/components/mockServer/CreateNewMockServerModal.vue
@@ -298,7 +298,7 @@ import { useReadonlyStream } from "@composables/stream"
 import { useToast } from "@composables/toast"
 import { computed, ref, watch } from "vue"
 import { TippyComponent } from "vue-tippy"
-import { MockServer } from "~/helpers/backend/graphql"
+import type { MockServer } from "~/helpers/backend/types/MockServer"
 import { showCreateMockServerModal$ } from "~/newstore/mockServers"
 import { useMockServer } from "~/composables/useMockServer"
 import MockServerCreatedInfo from "~/components/mockServer/MockServerCreatedInfo.vue"
